move analytical evolution to circuitprocessor
