+++ qutip/qip/models/modelprocessor.py
@@ -167,112 +167,6 @@ class ModelProcessor(CircuitProcessor):
         """
         return (self.ctrls, self.coeffs.T)
 
-    def run(self, qc=None):
-        """
-        Calculate the propagator the evolution by matrix exponentiation.
-        This method won't include the noise.
-
-        Parameters
-        ----------
-        qc: :class:`qutip.qip.QubitCircuit`, optional
-            Takes the quantum circuit to be implemented. If not given, use
-            the quantum circuit saved in the processor by `load_circuit`.
-
-        Returns
-        --------
-        U_list: list
-            The propagator matrix obtained from the physical implementation.
-        """
-        if qc:
-            self.load_circuit(qc)
-
-        U_list = []
-        tlist = self.tlist
-        for n in range(len(tlist)-1):
-            H = sum([self.coeffs[m, n] * self.ctrls[m]
-                    for m in range(len(self.ctrls))])
-            dt = tlist[n+1] - tlist[n]
-            U = (-1j * H * dt).expm()
-            U = self.eliminate_auxillary_modes(U)
-            U_list.append(U)
-
-        if self.correct_global_phase and self.global_phase != 0:
-            U_list.append(globalphase(self.global_phase, N=self.N))
-
-        return U_list
-
-    def run_state(self, qc=None, rho0=None, states=None,
-                  numerical=False, **kwargs):
-        """
-        Simulate the state evolution under the given `qutip.QubitCircuit`
-        If `numerical` is False, it will calculate the propagator
-        with matrix exponentiation and applies it to the initial state.
-        This won't includes the noise.
-        If `numerical` is True, it will use mesolve to
-        calculate the time of the state evolution
-        and return the result. Other arguments of mesolve can be
-        given as kwargs.
-
-        Parameters
-        ----------
-        qc: :class:`qutip.qip.QubitCircuit`, optional
-            Takes the quantum circuit to be implemented. If not given, use
-            the quantum circuit saved in the processor by `load_circuit`.
-
-        rho0: :class:`qutip.Qobj`, optional
-            Initial state of the qubits in the register.
-            If qutip solver is used to calculate the evolution numerically.
-            Noise will only be considered if this is true.
-    
-        kwargs, optional
-            Key word arguments for the:func:`qutip.mesolve`.
-    
-        states: :class:`qutip.Qobj`, optional
-            Old API, same as rho0.
-
-        Returns
-        --------
-        evo_result: :class:`qutip.Result`
-            If the numerical method is used, the result of the solver will
-            be returned.
-            An instance of the class :class:`qutip.Result`, which contains
-            either an *array* `result.expect` of expectation values for
-            the times specified by `tlist`, or an *array* `result.states`
-            of state vectors or density matrices corresponding to
-            the times in `tlist` [if `e_ops` is
-            an empty list], or nothing if a callback function was
-            given in place of
-            operators for which to calculate the expectation values.
-
-        U_list: list
-            If the analytical method is used, the propagator is returned.
-        """
-        if rho0 is None and states is None:
-            raise ValueError("Qubit state not defined.")
-        elif rho0 is None:
-            # just to keep the old prameters `states`, it is replaced by rho0
-            rho0 = states
-        if qc:
-            self.load_circuit(qc)
-            # TODO If noise exists, give warning
-        if numerical:
-            return super(ModelProcessor, self).run_state(rho0=rho0, **kwargs)
-
-        U_list = [rho0]
-        tlist = self.tlist
-        for n in range(len(tlist)-1):
-            H = sum([self.coeffs[m, n] * self.ctrls[m]
-                    for m in range(len(self.ctrls))])
-            dt = tlist[n+1] - tlist[n]
-            U = (-1j * H * dt).expm()
-            U = self.eliminate_auxillary_modes(U)
-            U_list.append(U)
-
-        if self.correct_global_phase and self.global_phase != 0:
-            U_list.append(globalphase(self.global_phase, N=self.N))
-
-        return U_list
-
     def pulse_matrix(self):
         """
         Generates the pulse matrix for the desired physical system.
